CogVM source as per VMMaker.oscogSPC-eem.2100
Temporary branch for testing the SpurPlanningCompactor.
Otherwise the same as VMMaker.oscog-eem.2099.

StackToRegisterMappingCogits:
Follow jumps to jumps and push: aBoolean; jump:if:s, eliminating dead code.
Elimnate jumps to the immediately following instruction in
StackToRegisterMappingCogit>>generateInstructionsAt:.
Simplify the various gen*InlinedIdenticalOrNotIf: to eliminate the duplication
using #== to compare orNot with the branch.

Sista:
Added support for inlined primitive #characterAsInteger.

Lowcode:
I keep working on the 64 bits version of Lowcode. Now I started writing the actual
ABI compiler and I got most of the tests for the UFFI passing (except for a time
to time crash on the qsort test), with both the Cogit and the StackInterpreter.
All of the tests for WodenMath are passing.