CogVM source as per VMMaker.oscogSPC-eem.2132
Temporary branch for the SpurPlanningCompactor.

SpurPlanningCompactor:
Fix the single-pass unmarking pass.  It needs to begin at firstFreeObject, not
lastMobileObject!!

Spur:
Provide isUnmarked: for debugging with printOopsSuchThat:.
Remember the first unmarked object found by allObjectsUnmarked in the bogon
variable, and keep hold on to it via printBogons so it is not deleted by the
C optimizer.

Fix assert fails in Spur due to failing perform primitives not restoring the
function pointer and argumentCount correctly.

Cogit:
Have printCogMethods print totals of the number of methods & PICs in the code
zone.  Fix a slip in printCogMethodsWithPrimitive:.