CogVM source as per VMMaker.oscogSPC-eem.2134
StackInterpreter:
Fix serious and ancient bug with stack checking and event checking.  The event
check piggy backs off the stack check, but the old code would not handle a stack
overflow if there was an overflow and an event check at one and the same time,
allowing frames to grow down the stack page in cases where there are lots of
events (such as profiling).

SpurPlanningCompactor:
Fix bad slip in unmarkObjectsFromFirstFreeObject; it forgot to free the last
object in the sequence.

Cogit:
Fix checking the lists of cog methods (openPICList and unpairedMethodList) when
compacting the method zone.
Fix a bug with freeing a method that might have either a nextMethod or an IRC;
Must not free CogMethods.
Fix a couple of C compilation warnings.

With these changes the Cadence Glue External Tests run to completion without
assert fails on 64-bit linux.